move events annd dashboard to pages
