--- src/pages/Calendar.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Calendar = () => {
-  return (
-    <div>Calendar</div>
-  )
-}
-
-export default Calendar
